% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancetoland-package.R
\docType{data}
\name{distance_to_coastline_lowres}
\alias{distance_to_coastline_lowres}
\title{Distance to coast (low resolution)}
\description{
A `RasterLayer` layer with the distance (metres) to the Natural Earth
coastline. Scale 110, using the `rnaturalearth::ne_coastline` function.
}
