% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancetoland-package.R
\docType{package}
\name{distancetocoast-package}
\alias{distancetocoast}
\alias{distancetocoast-package}
\title{distancetocoast: Look Up Distance to Coastline}
\description{
Provides stored raster data to give approximate distance to the coast for anywhere in the world.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
