% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancetocoast-package.R
\docType{data}
\name{distance_to_coastline_lowres}
\alias{distance_to_coastline_lowres}
\alias{distance_to_coastline_50}
\alias{distance_to_coastline_10}
\title{Distance to coast (varying resolutions)}
\description{
All are \code{RasterLayer} layer with the distance (metres) to the Natural Earth
coastline.
\itemize{
\item \code{distance_to_coastline_lowres}  Scale 110  \code{rnaturalearth::ne_coastline(scale = 110)}.
\item \code{distance_to_coastline_50}  Scale 50 \code{rnaturalearth::ne_coastline(scale = 50)}.
\item \code{distance_to_coastline_10}  Scale 10  \code{rnaturalearth::ne_coastline(scale = 110)}.
}
}
